Drop compatibility with GNOME Shell 42 and bump version
The new workspace activation (#3) has been tested on GNOME 44 (by the PR
author) and GNOME 43 (by myself, since Arch still hasn’t upgraded), but
not on GNOME 42. I have no particular reason to expect that it won’t
there, but it still feels safer to leave those users on an older version
of the extension that’s definitely not broken, I think. If any of them
get in touch saying they’d like to have the workplace activation feature
and tested that it works with GNOME 42, I’ll be happy to add it back to
the metadata.json.