Declare compatibility with GNOME Shell 48 and bump version
According to the porting guide [1], get_window_actors() moved from Meta
to Meta.Compositor, but as far as I can tell it’s actually still
available on global, so we don’t need to change anything at the moment.
I added a comment just in case it needs to be fixed later.

[1]: https://gjs.guide/extensions/upgrading/gnome-shell-48.html#meta