Rearrange currentDesktopFirst handling code
With the main goal of doing less work in the no-sorting case (no need to
map *all* the actors to windows then, or to get the current worspace),
but it also makes the code shorter overall. Array.prototype.sort() is
guaranteed to be stable since 2019 (and has in practice been stable in
SpiderMonkey for *much* longer than that, I believe), so replacing the
array splicing code with “sort by 0” should be equivalent, if I’m not
mistaken.